requirements.txt
bcrypt
pystray
pillow
